requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
praw==7.7.1
python-dotenv==1.0.0
lxml==5.1.0
schedule==1.2.0